findRegex: replace xargs with our own buffering/loop
It may be a little less efficient, but this gives me me more control
over the grep process(es). Which will allow me to kill/interrupt it in a
later CL.

Also I should be able to use io.Pipe instead of os.Pipe, and not set
fds, which will look nicer.