Sourcegraph: basic support
A "sourcegraph" keyword is automatically added to the interface if we
were able to guess what repo the project corresponds to (which requires
a GOPATH set).
A 2-1 chord on the keyword will then open in the browser a search URL
query to the sourcegraph.com website.

Untested with -there flag.